feat(S1.2.3): add user profile managment endpoints
