--- backend/RideNow/src/main/java/rs/ac/uns/ftn/asd/ridenow/dto/profile/ChangePasswordRequestDTO.java
@@ -1,34 +1,0 @@
-package rs.ac.uns.ftn.asd.ridenow.dto.profile;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Min;
-
-
-public class ChangePasswordRequestDTO {
-
-    @NotBlank
-    private String currentPassword;
-
-    @NotBlank
-    private String newPassword;
-
-    @NotBlank
-    private String confirmNewPassword;
-
-    public ChangePasswordRequestDTO() { super(); }
-
-    public String getCurrentPassword() { return currentPassword; }
-
-    public void setCurrentPassword(String currentPassword) { this.currentPassword = currentPassword; }
-
-    public String getNewPassword() { return newPassword; }
-
-    public void setNewPassword(String newPassword) { this.newPassword = newPassword; }
-
-    public String getConfirmNewPassword() { return confirmNewPassword; }
-
-    public void setConfirmNewPassword(String confirmNewPassword) { this.confirmNewPassword = confirmNewPassword; }
-
-}
-
