--- backend/RideNow/src/main/java/rs/ac/uns/ftn/asd/ridenow/dto/profile/UpdateProfileRequestDTO.java
@@ -1,88 +1,0 @@
-package rs.ac.uns.ftn.asd.ridenow.dto.profile;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Min;
-import rs.ac.uns.ftn.asd.ridenow.model.enums.VehicleType;
-
-
-public class UpdateProfileRequestDTO {
-
-    @Email
-    @NotBlank
-    private String email;
-
-    @NotBlank
-    private String firstName;
-
-    @NotBlank
-    private String lastName;
-
-    @NotBlank
-    private String phoneNumber;
-
-    @NotBlank
-    private String address;
-
-    private String profileImage;
-
-    @NotBlank
-    private String vehicleModel;
-
-    @Min(1)
-    private int numberOfSeats;
-
-    @NotNull
-    private VehicleType vehicleType;
-
-    private Boolean babyFriendly;
-    private Boolean petFriendly;
-
-    public UpdateProfileRequestDTO() { super(); }
-
-    public String getEmail() { return email; }
-
-    public void setEmail(String email) { this.email = email; }
-
-    public String getFirstName() { return firstName; }
-
-    public void setFirstName(String firstName) { this.firstName = firstName; }
-
-    public String getLastName() { return lastName; }
-
-    public void setLastName(String lastName) { this.lastName = lastName; }
-
-    public String getPhoneNumber() { return phoneNumber; }
-
-    public void setPhoneNumber(String phoneNumber) { this.phoneNumber = phoneNumber; }
-
-    public String getAddress() { return address; }
-
-    public void setAddress(String address) { this.address = address; }
-
-    public String getProfileImage() { return profileImage; }
-
-    public void setProfileImage(String profileImage) { this.profileImage = profileImage; }
-
-    public String getVehicleModel() { return vehicleModel; }
-
-    public void setVehicleModel(String vehicleModel) { this.vehicleModel = vehicleModel; }
-
-    public int getNumberOfSeats() { return numberOfSeats; }
-
-    public void setNumberOfSeats(int numberOfSeats) { this.numberOfSeats = numberOfSeats; }
-
-    public VehicleType getVehicleType() { return vehicleType; }
-
-    public void setVehicleType(VehicleType vehicleType) { this.vehicleType = vehicleType; }
-
-    public Boolean getBabyFriendly() { return babyFriendly; }
-
-    public void setBabyFriendly(Boolean babyFriendly) { this.babyFriendly = babyFriendly; }
-
-    public Boolean getPetFriendly() { return petFriendly; }
-
-    public void setPetFriendly(Boolean petFriendly) { this.petFriendly = petFriendly; }
-    
-}
-
