+++ backend/RideNow/src/main/java/rs/ac/uns/ftn/asd/ridenow/dto/user/ChangePasswordRequestDTO.java
@@ -1,0 +1,23 @@
+package rs.ac.uns.ftn.asd.ridenow.dto.user;
+import jakarta.validation.constraints.NotBlank;
+import lombok.Getter;
+import lombok.Setter;
+
+
+@Setter
+@Getter
+public class ChangePasswordRequestDTO {
+
+    @NotBlank
+    private String currentPassword;
+
+    @NotBlank
+    private String newPassword;
+
+    @NotBlank
+    private String confirmNewPassword;
+
+    public ChangePasswordRequestDTO() { super(); }
+
+}
+
